+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🎉 feat: handler generation added
 - 🎉 feat: added Me route and expanded repositories generation part
 - 🎉 feat: added login with phone and login with email address
 - 🐛 fix: bug fix on translator and query generation